+++ NEWS.md
@@ -8,6 +8,8 @@ Install the development version with: `install_github("rstudio/reticulate")`
 - `py_dict()` function to enable creation of dictionaries based on lists of 
   keys and values.
 
+- Provide default base directory (e.g. '~/.virtualenvs') for environments
+  specified by name in `use_virtualenv()`.
 
 ## reticulate 1.5 (CRAN)