+++ NEWS.md
@@ -5,6 +5,10 @@ Install the development version with: `install_github("rstudio/reticulate")`
 
 - `py_repl()` function implementing a lightweight Python REPL in R.
 
+- Support for converting Pandas objects (`Index`, `Series`, `DataFrame`)
+
+- Support for converting Python `datetime` objects.
+
 - `py_dict()` function to enable creation of dictionaries based on lists of 
   keys and values.