+++ NEWS.md
@@ -3,7 +3,7 @@
 
 Install the development version with: `install_github("rstudio/reticulate")`
 
-- `py_repl()` function implementing a lightweight Python REPL in R.
+- `repl_python()` function implementing a lightweight Python REPL in R.
 
 - Support for converting Pandas objects (`Index`, `Series`, `DataFrame`)